scripts/ingestion/requirements.txt
# Python dependencies for documentation ingestion scripts

# Database
asyncpg>=0.29.0

# HTTP requests and async
aiohttp>=3.9.0
requests>=2.31.0

# HTML parsing and data processing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# OpenAI API
openai>=1.7.0

# Utilities
python-dotenv>=1.0.0